% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtest.R
\name{regtest}
\alias{regtest}
\title{test if a atomic vector endwith "le"}
\usage{
regtest(s)
}
\arguments{
\item{s}{an character atomic vector}
}
\value{
logical
}
\description{
test if a atomic vector endwith "le"
}
\examples{
regtest("apple")
}
